requirements.txt
asgiref==3.2.3
blessings==1.7
bpython==0.18
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
curtsies==0.3.1
Django==3.0.3
django-cors-headers==3.2.1
django-environ==0.4.5
django-environ-plus==0.0.3
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-composed-permissions==0.2.1
djangorestframework-role-filters==0.3.0
djangorestframework-simplejwt==4.4.0
greenlet==0.4.15
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
python-http-client==3.2.4
pytz==2019.3
requests==2.22.0
rest-condition==1.0.3
sendgrid==6.1.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
